+++ src/main/java/edu/architect_711/wordsapp/model/dto/account/AccountDetails.java
@@ -1,4 +1,4 @@
-package edu.architect_711.wordsapp.model.dto;
+package edu.architect_711.wordsapp.model.dto.account;
 
 import edu.architect_711.wordsapp.model.entity.Account;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,17 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Collection;
 import java.util.List;
 
-@AllArgsConstructor @NoArgsConstructor
+/**
+ * This object <b>MUST</b> be placed in the {@link org.springframework.security.core.Authentication} object
+ * as the principal, because it saves from the extra SQL-request to the database, for example to get the
+ * authenticated user id.
+ * <p>
+ * This is very helpful. You don't need to search for the authenticated user entity in the database,
+ * because it's already in the Authentication, since a user went through the {@link org.springframework.security.web.SecurityFilterChain}
+ * and loaded from the database by the {@link edu.architect_711.wordsapp.service.account.AccountDetailsService}
+ */
+@AllArgsConstructor
+@NoArgsConstructor
 @Data
 public class AccountDetails implements UserDetails {
     private Account account;
