+++ src/main/java/edu/architect_711/wordsapp/model/dto/account/AccountDto.java
@@ -1,4 +1,4 @@
-package edu.architect_711.wordsapp.model.dto;
+package edu.architect_711.wordsapp.model.dto.account;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -7,8 +7,17 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor
+/**
+ * This object is usually used to return the saved
+ * {@link edu.architect_711.wordsapp.model.entity.Account}.
+ * <p>
+ * The `password` field is absent for the security terms
+ * (and why if it's encrypted?)
+ */
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class AccountDto {
     @NotNull
     private Long id;
