+++ src/main/java/edu/architect_711/wordsapp/model/dto/group/UpdateGroupDto.java
@@ -1,29 +1,17 @@
-package edu.architect_711.wordsapp.model.dto;
+package edu.architect_711.wordsapp.model.dto.group;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
-@AllArgsConstructor @NoArgsConstructor
-@Data
-public class GroupDto {
+@Data @AllArgsConstructor
+public class UpdateGroupDto {
     @NotNull @Min(0)
     private Long id;
-
     @NotBlank
     private String title;
-
-    @NotNull
-    private LocalDateTime created;
-
     @NotBlank
     private String description;
-
-    @NotNull @Min(0)
-    private Long ownerId;
 }
